テストケースをある程度実装

+++ src/main/java/com/example/webapp/service/impl/AdminServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.example.webapp.entity.Authentication;
+import com.example.webapp.exception.WebappException;
 import com.example.webapp.repository.AuthenticationMapper;
 import com.example.webapp.service.AdminService;
 import com.example.webapp.utility.PasswordGenerator;
@@ -20,28 +21,40 @@ public class AdminServiceImpl implements AdminService {
 
 	@Override
 	public List<Authentication> getUserList() {
-		return authenticationMapper.selectUserList();
+		List<Authentication> resultList = authenticationMapper.selectUserList();
+		if (resultList == null) {
+			throw new WebappException("ユーザーが存在しません");
+		}
+		return resultList;
 	}
 
 	@Override
-	public void registUser(Authentication auth) {
+	public String registUser(Authentication auth) {
 		// パスワードをハッシュ化
 		String hashedPassword = PasswordGenerator.generateHashedPassword(auth.getPassword());
 		auth.setPassword(hashedPassword);
 
 		authenticationMapper.insertUser(auth);
+		return "ユーザー：" + auth.getUsername() + " を登録しました。";
 	}
 
 	@Override
-	public void deleteUser(String username) {
+	public String deleteUser(String username) {
+		// 削除対象が存在するかチェック
+		Authentication auth = authenticationMapper.selectByUsername(username);
+		if (auth == null) {
+			throw new WebappException("削除対象が存在しません");
+		}
+
 		authenticationMapper.deleteUser(username);
+		return "ユーザー：" + username + " 削除しました";
 	}
 
 	@Override
-	public boolean isNotDeleteUser(String username) {
+	public void isNotDeleteUser(String username) {
 		// ユーザー「admin」の場合NG
 		if (username.equals("admin")) {
-			return true;
+			throw new WebappException("adminは削除できません");
 		}
 		// カレントユーザー名の取得
 		org.springframework.security.core.Authentication authentication = SecurityContextHolder.getContext()
@@ -50,26 +63,17 @@ public class AdminServiceImpl implements AdminService {
 
 		// カレントユーザーの場合NG
 		if (username.equals(currentUsername)) {
-			return true;
+			throw new WebappException("操作中のユーザーは削除できません");
 		}
-		return false;
 	}
 
 	@Override
-	public boolean isRegistUser(String username) {
+	public void isRegistUser(String username) {
 		// 登録ユーザーと同一のユーザーがいないか確認
 		Authentication auth = authenticationMapper.selectByUsername(username);
 		// 重複無し
 		if (auth == null) {
-			return false;
+			throw new WebappException("既に使われているユーザー名です");
 		}
-		return true;
 	}
-
-	@Override
-	public String createHashedPassword(String warPassword) {
-		// TODO 自動生成されたメソッド・スタブ
-		return null;
-	}
-
 }
